メインコンポーネントを切り出す
