+++ app/javascript/packs/setting_editor.tsx
@@ -4,53 +4,22 @@ import { jsx, css } from '@emotion/react'
 
 import React, { useState } from 'react'
 import ReactDOM from 'react-dom'
-import PropTypes from 'prop-types'
-import { Button, buttons } from "types/button";
-import { ModalProps, ButtonsModal } from "./setting_editor/components/buttons_modal";
-import { InstallableModes } from "./setting_editor/components/installable_modes";
-import { useModal } from "./setting_editor/hooks/use_modal";
+import { ButtonsSetting } from './setting_editor/components/main'
+import { Button } from "types/button";
 
-type Props = {
-  defaultPrefixKeys: Array<Button>;
-};
-
-export const ButtonsSettingPage = ({ defaultPrefixKeys }) => {
-  const [modalProps, openModal] = useModal();
-  const [prefixKeys, setPrefixKeys] = useState(defaultPrefixKeys);
-  const handlePrefixKeysField = () => {
-    openModal({ title: "キープレフィックスの変更", prefill: prefixKeys, callbackOnSubmit: setPrefixKeys });
-  }
+export const ButtonsSettingPage = () => {
+  const prefixKeys = JSON.parse(document.getElementById("config-prefix-keys").dataset.configPrefixKeys) as Array<Button>
 
   return(
     <>
-      <div css={css`display: table`}>
-        <div css={css`display: table-cell; width: 400px;`}>
-          <h2>設定ファイルの変更</h2>
-          <div>
-            <a href="#" >エクスポートする</a>
-          </div>
-
-          <h3>インストール可能なモード</h3>
-          {<InstallableModes />}
-
-          <h3>インストール可能なマクロ</h3>
-
-          <h3>設定中のプレフィックスキー</h3>
-          <input type="text" value={prefixKeys.join(", ")} readOnly={true} onClick={handlePrefixKeysField} />
-          {<ButtonsModal {...modalProps as ModalProps} />}
-        </div>
-        <div css={css`display: table-cell`}>
-        </div>
-      </div>
+      <ButtonsSetting defaultPrefixKeys={prefixKeys} />
     </>
   )
 }
 
-const prefixKeys = JSON.parse(document.getElementById("config-prefix-keys").dataset.configPrefixKeys) as Array<Button>
-
 document.addEventListener('DOMContentLoaded', () => {
   ReactDOM.render(
-    <ButtonsSettingPage defaultPrefixKeys={prefixKeys} />,
+    <ButtonsSettingPage />,
     document.body.appendChild(document.createElement('div')),
   )
 })
