+++ app/javascript/packs/setting_editor/setting_context.ts
@@ -1,0 +1,2 @@
+import { createContext } from "react";
+export const SettingContext = createContext<any>(null);
